telegram_system/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
dnspython==2.2.1
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pymongo==4.0.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
Werkzeug==2.1.0
zipp==3.7.0
pandas
